refactor: move components and module.css files to components directory
