+++ src/components/newsBar/NewsBar.module.css
