+++ src/components/newsBar/NewsBar.ts
