+++ src/components/newsBar/RollingNews.ts
